Fix for ipa-client partial install issue
When ipa client is isntalled but not propperly configured ambari is not
able to install it anew as it gives an "already installed" error. Adding
a check if ipa client default conf file is present but the lock file we
create is not - remove the ipa client so it can be installed from the
beginning.